+++ metomi/rosie/svn_post_commit.py
@@ -45,7 +45,7 @@ from metomi.rose.resource import ResourceLocator
 from metomi.rose.scheme_handler import SchemeHandlersManager
 from metomi.rosie.db import (
     LATEST_TABLE_NAME, MAIN_TABLE_NAME, META_TABLE_NAME, OPTIONAL_TABLE_NAME)
-from metomi.rosie.svn_hook import RosieSvnHook
+from metomi.rosie.svn_hook import RosieSvnHook, InfoFileError
 
 
 class RosieWriteDAO(object):
@@ -195,9 +195,12 @@ class RosieSvnPostCommitHook(RosieSvnHook):
                 if tail == self.INFO_FILE:
                     # Suite info file change
                     if branch_attribs["info"] is None:
-                        branch_attribs["info"] = self._load_info(
-                            repos, sid, branch, revision=revision,
-                            allow_popen_err=True)
+                        try:
+                            branch_attribs["info"] = self._load_info(
+                                repos, sid, branch, revision=revision,
+                                allow_popen_err=True)
+                        except metomi.rose.config.ConfigSyntaxError as exc:
+                            raise InfoFileError(InfoFileError.VALUE, exc)
                     if path_status != self.ST_ADDED:
                         branch_attribs["old_info"] = self._load_info(
                             repos, sid, branch, revision=int(revision)-1,
@@ -218,9 +221,12 @@ class RosieSvnPostCommitHook(RosieSvnHook):
                     branch_attribs["status"] = path_status
                 # Load suite info and old info regardless
                 if branch_attribs["info"] is None:
-                    branch_attribs["info"] = self._load_info(
-                        repos, sid, branch, revision=revision,
-                        allow_popen_err=True)
+                    try:
+                        branch_attribs["info"] = self._load_info(
+                            repos, sid, branch, revision=revision,
+                            allow_popen_err=True)
+                    except metomi.rose.config.ConfigSyntaxError as exc:
+                        raise InfoFileError(InfoFileError.VALUE, exc)
                     # Note: if (allowed) popen err, no DB entry will be created
                 if (branch_attribs["old_info"] is None and
                         branch_attribs["status"] == self.ST_DELETED):
